extract constraintObjectRotation func
